intermediate-backend/M2_submission/NLRent/requirements.txt
Package             Version
------------------- ---------
asgiref             3.4.1
backports.zoneinfo  0.2.1
certifi             2021.10.8
charset-normalizer  2.0.9
Django              3.0.1
django-filter       21.1
djangorestframework 3.13.1
idna                3.3
importlib-metadata  4.10.0
Markdown            3.3.6
numpy               1.21.5
pandas              1.3.5
pip                 21.3.1
python-dateutil     2.8.2
pytz                2021.3
requests            2.26.0
setuptools          58.5.3
six                 1.16.0
sqlparse            0.4.2
urllib3             1.26.7
wheel               0.37.0
zipp                3.6.0
